Renamed UserServis`s methods. Unit tests for GetAllUsers method
